requirements.txt
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-api==1.27.0
litellm==1.48.7
python-dotenv==1.0.1
fastapi==0.115.2
uvicorn==0.30.6
httpx==0.27.2
